+++ JeanLF.AudioService.Editor/AudioServiceSettingsEditor.cs
@@ -5,6 +5,8 @@ using System.Linq;
 using UnityEditor;
 using UnityEditor.AddressableAssets;
 using UnityEditor.AddressableAssets.Settings;
+using UnityEditor.Build;
+using UnityEditor.Build.Reporting;
 using UnityEditor.UIElements;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
@@ -12,11 +14,29 @@ using UnityEngine.UIElements;
 
 namespace JeanLF.AudioService.Editor
 {
-    internal class AudioServiceSettingsEditor : AssetPostprocessor
+    internal class AudioServiceSettingsEditor : AssetPostprocessor, IPreprocessBuildWithReport
     {
         private static Button _editButton;
         private static SerializedObject _settings;
         private static ObjectField _configField;
+        
+        public int callbackOrder { get; }
+        
+        public void OnPreprocessBuild(BuildReport report)
+        {
+            AudioServiceSettings settings = GetOrCreateSettings();
+            Debug.Log($"Audio Service - Pre-build step\nGenerating code for {settings.Database}");
+
+            if (settings.Database != null)
+            {
+                if (!CheckOrCreateAddressableEntry(settings))
+                {
+                    Debug.LogError("Audio Service - Missing Addressable settings, Audio service will not work.");
+                    return;
+                }
+                GenerateCode(settings.Database);
+            }
+        }
 
         private static void OnPostprocessAllAssets(string[] importedAssets, string[] deletedAssets, string[] movedAssets, string[] movedFromAssetPaths)
         {
@@ -58,18 +78,28 @@ namespace JeanLF.AudioService.Editor
                 AssetDatabase.SaveAssets();
             }
 
+            CheckOrCreateAddressableEntry(settings);
+            return settings;
+        }
+
+        private static bool CheckOrCreateAddressableEntry(AudioServiceSettings settings)
+        {
+            if (!AddressableAssetSettingsDefaultObject.SettingsExists)
+            {
+                return false;
+            }
+            
             string guid = AssetDatabase.AssetPathToGUID(AudioServiceEditorUtils.SettingsAssetPath);
 
             if (AddressableAssetSettingsDefaultObject.Settings.FindAssetEntry(guid) != null)
             {
-                return settings;
+                return true;
             }
 
             AddressableAssetEntry entry = AddressableAssetSettingsDefaultObject.Settings.CreateOrMoveEntry(guid, AddressableAssetSettingsDefaultObject.Settings.DefaultGroup);
             entry.address = AudioServiceSettings.FileName;
             AddressableAssetSettingsDefaultObject.Settings.SetDirty(AddressableAssetSettings.ModificationEvent.EntryMoved, entry, true);
-
-            return settings;
+            return true;
         }
 
         internal static void GenerateCode(AudioDatabase database)
@@ -87,7 +117,8 @@ namespace JeanLF.AudioService.Editor
 
         private static void DrawSettings(string searchContext, VisualElement rootElement)
         {
-            _settings = new SerializedObject(GetOrCreateSettings());
+            AudioServiceSettings audioSettings = GetOrCreateSettings();
+            _settings = new SerializedObject(audioSettings);
             SerializedProperty configProp = _settings.FindProperty(AudioServiceSettings.DatabaseName);
 
             VisualTreeAsset visualTree =
@@ -119,6 +150,13 @@ namespace JeanLF.AudioService.Editor
 
             PropertyField shrinkField = rootElement.Q<PropertyField>("shrinkCount");
             shrinkField.BindProperty(poolProp.FindPropertyRelative(PoolSettings.ShrinkCountName));
+
+            if (!CheckOrCreateAddressableEntry(audioSettings))
+            {
+                var helpBox = new HelpBox("Missing Unity's addressable settings, Audio service needs it to work!\nOpen 'Window/Asset Management/Addressables/Groups' for more info.",
+                    HelpBoxMessageType.Warning);
+                rootElement.Add(helpBox);
+            }
         }
 
         private static void OnConfigChange(ChangeEvent<UnityEngine.Object> evt)