+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## [1.1.0] - 2024-04-12
+- Add an event for when the AudioPlayer change its track;
+- Add an event for when the AudioPlayer ends its execution;
+- Add a warning if the project doesn't have Addressables configured correctly.
+
 ## [1.0.2] - 2023-01-31
 - Fix error with tempo events
 - Add a Pre Build step to garantee code generation